+++ R/qmhn.R
@@ -49,7 +49,7 @@
 #' Sun, J., Kong, M., & Pal, S. (2023). The Modified-Half-Normal
 #' distribution: Properties and an efficient sampling scheme.
 #' \emph{Communications in Statistics - Theory and Methods}, 52(5),
-#' 1507--1536.
+#' 1591--1613.
 #'
 #' @seealso \code{\link{dmhn}}, \code{\link{pmhn}}, \code{\link{rmhn}}
 #'