+++ R/rmhn.R
@@ -69,7 +69,7 @@
 #' Sun, J., Kong, M., & Pal, S. (2023). The Modified-Half-Normal
 #' distribution: Properties and an efficient sampling scheme.
 #' \emph{Communications in Statistics - Theory and Methods}, 52(5),
-#' 1507--1536.
+#' 1591--1613.
 #'
 #' Gao, F. & Wang, H.-B. (2025). Generating modified-half-normal random
 #' variates by a relaxed transformed density rejection method.